Docstring for Induction Motor Circle and some more changes
