+++ .gitignore
@@ -8,6 +8,7 @@ docs/
 dist/
 build/
 *.egg-info/
+.idea/
 
 # Ignore TOX
 .tox/
