+++ electricpy/__init__.py
@@ -6511,7 +6511,7 @@ def phs3valpha(VA, VB=0, VC=0):
     """
     # Handle Combined (list/tuple) Input
     if (isinstance(VA, (tuple, list)) and
-            VB == 0 and VC == C):
+            VB == 0 and VC == 0):
         if len(VA) != 3:
             raise ValueError("Invalid input set, must "
                              + "be list of three elements, three inputs,"
@@ -6573,29 +6573,64 @@ def wireresistance(length=None,diameter=None,rho=16.8*10**-9,R=None):
         A = rho*length/R
         return _np.sqrt(4*A/pi)
         
-def ic_555(mode,R=None,C=None,f=None,t_high=None,t_low=None):
+def ic_555(mode,R=None,C=None,freq=None,t_high=None,t_low=None):
+    """
+    555 Integrated Circuit Calculator 
+    
+    Evaluate a number of common attributes related to the common 555 integrated circuit including time period, frequency, duty cycle, time spent low during each cycle, time spent high during each cycle. 
+    
+    .. math:: <various-math-equations for each output> 
+    
+    .. image:: <a-diagram-showing-the-circuit-configuration-for-the-ic> 
+    
+    Parameters 
+    ---------- 
+    
+    mode: ['astable', 'monostable'] The operating constraint of the 555 timer. 
+    
+    R: list[float, float] or tuple(float, float) optional list of 2 resistor which are need in configuring IC 555
+    
+    C: float optional Capacitance between Threshold Pin and ground
+    
+    f: float optional Electrical system frequency in Hertz. 
+    
+    t_high: float, optional ON time of IC 555 
+    
+    t_low: float, optional OFF time of IC 555 
+    
+    Returns 
+    ------- 
+    
+    dict: "time_period": Time period of oscillating IC 555 
+    
+    "frequency": frequency of oscilation of IC 555 
+    
+    "duty_cycle": ration between ON time and total time
+    
+    "t_low": ON time of IC 555 
+    
+    "t_high": OFF time of IC 555 
+    """
 
     if mode == 'ASTABLE':
 
         if R!=None and C!=None:
 
-            try:
-                assert len(R) == 2
-            except AssertionError:
+            if len(R) != 2:
                 raise ValueError(f"ic_555 in {mode} will have only 2 resitances to be fixed but {len(R)} were \
                                 given")
 
             [R1, R2] = R
 
             T = _np.log(2)*C*(R1+2*R2)
-            f = 1/T
+            freq = 1/T
             t_low = _np.log(2)*C*R2
             t_high = _np.log(2)*C*(R1+R2)
             duty_cycle = t_high*100/T
 
             return {
                 'time_period':T,
-                'frequency':f,
+                'frequency':freq,
                 'duty_cycle':duty_cycle,
                 't_low':t_low,
                 't_high':t_high
@@ -6608,13 +6643,13 @@ def ic_555(mode,R=None,C=None,f=None,t_high=None,t_low=None):
 
             T = t_high+t_low
 
-            f = 1/(T)
+            freq = 1/(T)
 
             duty_cycle = t_high/(T)
 
             return {
                 'time_period':T,
-                'frequency':f,
+                'frequency':freq,
                 'duty_cycle':duty_cycle,
                 'R1':x1-x2,
                 'R2':x2
@@ -6647,203 +6682,4 @@ def ic_555(mode,R=None,C=None,f=None,t_high=None,t_low=None):
 
     else:
         raise ValueError("ic_555 is configured only in `MONOSTABLE` and `ASTABLE` modes")
-
-        
-class InductionMotorCircle:
-    def __init__(self, no_load_data, blocked_rotor_data, output_power, torque_ration=1, frequency=50, poles=4):
-
-        self.no_load_data = no_load_data
-        self.blocked_rotor_data = blocked_rotor_data
-        self.f = frequency
-        self.operating_power = output_power
-        self.torque_ratio = torque_ration
-        self.poles = poles
-        self.sync_speed = 120*frequency/poles #rpm
-
-        v0 = no_load_data['V0']
-        i0 = no_load_data['I0']
-        w0 = no_load_data['W0']
-
-        self.no_load_pf = w0 / (_np.sqrt(3) * v0 * i0)
-        theta0 = _np.arccos(self.no_load_pf)
-
-        # get short circuit power factor and Current at slip=1
-        vsc = blocked_rotor_data['Vsc']
-        isc = blocked_rotor_data['Isc']
-        wsc = blocked_rotor_data['Wsc']
-
-        self.blocked_rotor_pf = wsc / (_np.sqrt(3) * vsc * isc)
-        theta_sc = _np.arccos(self.blocked_rotor_pf)
-
-        # because V is on Y axis
-        theta0 = _np.pi / 2 - theta0
-        theta_sc = _np.pi / 2 - theta_sc
-
-        # isc is the current at reduced voltage so calculate current at rated voltage
-        isc = v0 * isc / vsc
-        self.no_load_line = [[0, i0 * _np.cos(theta0)], [0, i0 * _np.sin(theta0)]]
-        self.full_load_line = [[0, isc * _np.cos(theta_sc)], [0, isc * _np.sin(theta_sc)]]
-
-        # secondary current line
-        self.secondary_current_line = [[i0 * _np.cos(theta0), isc * _np.cos(theta_sc)],
-                                       [i0 * _np.sin(theta0), isc * _np.sin(theta_sc)]]
-
-        [[x1, x2], [y1, y2]] = self.secondary_current_line
-        self.theta = _np.arctan((y2 - y1) / (x2 - x1))
-
-        # get the induction motor circle
-        self.power_scale = w0 / (i0 * _np.sin(theta0))
-        self.center, self.radius = self.compute_circle_params()
-        [self.center_x, self.center_y] = self.center
-        self.p_max = self.radius * _np.cos(self.theta) - (self.radius - self.radius * _np.sin(self.theta)) * _np.tan(self.theta)
-        self.torque_line, self.torque_point = self.get_torque_line()
-        self.torque_max, self.torque_max_x, self.torque_max_y = self.get_torque_max()
-        _, [self.power_x, self.power_y] = self.get_output_power()  # Take low slip point
-        self.data = self.compute_efficiency()
-
-    def __call__(self):
-        return self.data
-
-    def plot(self):
-        [circle_x, circle_y] = InductionMotorCircle.get_circle(self.center, self.radius, semi=True)
-        _plt.plot(circle_x, circle_y)
-
-        InductionMotorCircle.plot_line(self.no_load_line)
-        InductionMotorCircle.plot_line(self.secondary_current_line)
-        InductionMotorCircle.plot_line(self.full_load_line, ls='-.')
-        InductionMotorCircle.plot_line(self.torque_line, ls='-.')
-
-        _plt.plot([self.secondary_current_line[0][1], self.secondary_current_line[0][1]],
-                 [self.secondary_current_line[1][1], self.center_y])  # Full load output
-        _plt.plot([self.center_x - self.radius, self.center_x+self.radius],
-                 [self.center_y, self.center_y], ls='-.')  # Diameter of the circle
-        _plt.plot([self.center_x, self.torque_max_x], [self.center_y, self.torque_max_y], ls='-.')  # Max torque line
-        _plt.plot([self.center_x, self.center_x - self.radius * _np.sin(self.theta)],
-                 [self.center_y, self.center_y + self.radius * _np.cos(self.theta)], ls='-.')  # Max Output Power line
-        _plt.plot([0, self.power_x], [0, self.power_y], c='black')  # Operating Point
-
-
-        _plt.scatter(self.power_x, self.power_y, marker='X', c='red')
-        _plt.scatter(self.center_x, self.center_y, marker='*', c='blue')  # mark the center of the circle
-        _plt.scatter(self.center_x - self.radius * _np.sin(self.theta), self.center_y + self.radius * _np.cos(self.theta),
-                    linewidths=3, c='black', marker='*')
-        _plt.scatter(self.torque_max_x, self.torque_max_y,
-                    linewidths=3, c='black', marker='*')
-
-
-        _plt.title("Induction Motor Circle Diagram")
-        _plt.grid()
-        _plt.legend(['I2 locus', 'No Load Current', 'Output Line', 'Blocked Rotor Current', 'Torque line',
-                    'Full Load Losses', 'Diameter', 'Maximum Torque',
-                    'Maximum Output Power', f'Operating Power {self.operating_power}'])
-        _plt.show()
-
-    def compute_efficiency(self):
-        [[_, no_load_x], [_, no_load_y]] = self.no_load_line
-        no_load_losses = no_load_y * self.power_scale
-
-        compute_slope = InductionMotorCircle.compute_slope
-
-        torque_slope = compute_slope(self.torque_line)
-        stator_cu_loss = (self.power_x - no_load_x) * torque_slope * self.power_scale
-
-        rotor_current_slope = compute_slope(self.secondary_current_line)
-        total_cu_loss = (self.power_x - no_load_x) * rotor_current_slope * self.power_scale
-
-        rotor_cu_loss = total_cu_loss - stator_cu_loss
-
-        rotor_output = self.power_y * self.power_scale - (rotor_cu_loss + stator_cu_loss + no_load_losses)
-
-        slip = rotor_cu_loss / rotor_output
-
-        self.rotor_speed = self.sync_speed*(1-slip)
-
-        data = {
-            'no_load_loss': no_load_losses,
-            'rotor_copper_loss': rotor_cu_loss,
-            'stator_copper_loss': stator_cu_loss,
-            'rotor_output': rotor_output,
-            'slip': slip,
-            'stator_rmf_speed (RPM)':self.sync_speed,
-            'rotor_speed (RMP)':self.rotor_speed,
-            'power_factor': (self.power_y / _np.sqrt(self.power_x ** 2 + self.power_y ** 2)),
-            'efficiency': f"{rotor_output * 100 / (self.power_y * self.power_scale)} %"
-        }
-        return data
-
-    @staticmethod
-    def get_circle(center, radius, semi=False):
-
-        [x0, y0] = center
-
-        if semi:
-            theta = _np.arange(0, _np.pi, 1e-4)
-        else:
-            theta = _np.arange(0, _np.pi * 2, 1e-4)
-
-        x = x0 + radius * _np.cos(theta)
-        y = y0 + radius * _np.sin(theta)
-        return x, y
-
-    @staticmethod
-    def plot_line(line, mark_start=True, mark_end=True, ls='-', marker=None):
-        [x, y] = line
-        [x1, x2] = x
-        [y1, y2] = y
-        _plt.plot(x, y, ls=ls)
-        if mark_start:
-            _plt.scatter(x1, y1, marker=marker)
-        if mark_end:
-            _plt.scatter(x2, y2, marker=marker)
-
-    def compute_circle_params(self):
-        [[x1, x2], [y1, y2]] = self.secondary_current_line
-        theta = _np.arctan((y2 - y1) / (x2 - x1))
-        length = _np.sqrt((x2 - x1) ** 2 + (y2 - y1) ** 2)
-        radius = length / (2 * _np.cos(theta))
-        center = [radius + x1, y1]
-
-        return center, radius
-
-    def get_torque_line(self):
-        [[x1, x2], [y1, y2]] = self.secondary_current_line
-        y = (self.torque_ratio * y2 + y1) / (self.torque_ratio + 1)
-        torque_point = [x2, y]
-        torque_line = [[x1, x2], [y1, y]]
-        return torque_line, torque_point
-
-    def get_torque_max(self):
-        [x, y] = self.torque_line
-        [x1, x2] = x
-        [y1, y2] = y
-        alpha = _np.arctan((y2 - y1) / (x2 - x1))
-        torque_max = self.radius * _np.cos(alpha) - (self.radius - self.radius * _np.sin(alpha)) * _np.tan(alpha)
-        torque_max_x = self.center_x - self.radius * _np.sin(alpha)
-        torque_max_y = self.center_y + self.radius * _np.cos(alpha)
-        return torque_max, torque_max_x, torque_max_y
-
-    @staticmethod
-    def compute_slope(line):
-        [[x1, x2], [y1, y2]] = line
-        return (y2 - y1)/(x2 - x1)
-
-    def get_output_power(self):
-        [[x1, x2], [y1, y2]] = self.secondary_current_line
-        alpha = _np.arctan((y2 - y1) / (x2 - x1))
-        [center_x, center_y] = self.center
-
-        [[_, no_load_x], [_, _]] = self.no_load_line
-        beta = _np.arcsin(
-            (self.operating_power / self.power_scale + (center_x - no_load_x) * _np.tan(alpha)) *
-            _np.cos(alpha) / self.radius)
-        beta_0 = alpha + beta
-        beta_1 = -alpha + beta
-        # high slip
-        p_x_1 = center_x + self.radius * _np.cos(beta_0)
-        p_y_1 = center_y + self.radius * _np.sin(beta_0)
-        # low slip
-        p_x_2 = center_x - self.radius * _np.cos(beta_1)
-        p_y_2 = center_y + self.radius * _np.sin(beta_1)
-        return [p_x_1, p_y_1], [p_x_2, p_y_2]    
-
 # END OF FILE
