+++ test/test_electricpy.py
@@ -243,7 +243,7 @@ def test_ic_555():
     #test the other way around
 
 def test_induction_motor_circle():
-    from electricpy import InductionMotorCircle
+    from electricpy.InductionMotorCircle import InductionMotorCircle
 
     open_circuit_test_data = {'V0': 400, 'I0': 9, 'W0': 1310}
     blocked_rotor_test_data = {'Vsc': 200, 'Isc': 50, 'Wsc': 7100}
